+++ languages/src/lib.rs
@@ -11,7 +11,6 @@
 
 pub mod ambient;
 pub mod calculator;
-pub(crate) mod for_clause;
 pub mod lambda;
 pub mod rhocalc;
