+++ languages/src/rhocalc.rs
@@ -5,9 +5,11 @@
 )]
 
 use mettail_macros::language;
-use mettail_runtime::{Language, Term, TermType, VarTypeInfo};
 use num_traits::Zero;
 use std::ops::Neg;
+
+pub(crate) mod receive;
+mod type_inference;
 
 language! {
     name: RhoCalc,
@@ -87,7 +89,7 @@ language! {
         // Internal guard gate used by where-clause gating.
         GuardThen . cond:Proc, body:Proc
         |- "__guard_then" "(" cond "," body ")" : Proc ![{
-            crate::for_clause::guard_then(&cond, &body)
+            crate::rhocalc::receive::guard_then(&cond, &body)
         }] fold;
 
         // Internal helper for where-guarded communication.
@@ -95,7 +97,7 @@ language! {
         // receive/send pair unchanged (blocked communication, identity).
         CommWhere . pat:Proc, n:Name, q:Proc, cond:Proc, body:Proc
         |- "__comm_where" "(" pat "<-" n "," q "," cond "," body ")" : Proc ![{
-            crate::for_clause::comm_pforwhere_subst(&pat, &n, &q, &cond, &body)
+            crate::rhocalc::receive::comm_pforwhere_subst(&pat, &n, &q, &cond, &body)
         }] fold;
 
         // Single pattern/channel binding: `pat <- chan`.
@@ -125,7 +127,7 @@ language! {
         // can contain `&`-joined binds with optional `where`.
         PForUser . rows:Vec(ForRow), body:Proc
         |- "for" "(" rows.*sep(";") ")" "{" body "}" : Proc ![{
-            crate::for_clause::desugar_for_rows(rows, body)
+            crate::rhocalc::receive::desugar_for_rows(rows, body)
         }] fold;
 
 
@@ -882,7 +884,7 @@ language! {
             |- (PPar {(PForWhere pat n cond body), (POutput n q), ...rest})
             ~> (PPar {(CommWhere pat n q cond body), ...rest});
 
-        // Zip+Map codegen derives `ns` from `b`/`bs` (see `pattern.rs` + `for_clause::channel_names_from_row`).
+        // Zip+Map codegen derives `ns` from `b`/`bs` (see `pattern.rs` + `receive::channel_names_from_row`).
         Comm . |- (PPar {(PForJoin b bs cond body), *zip(ns,qs).*map(|n,q| (POutput n q)), ...rest})
             ~> (PPar {(comm_join b bs ns qs cond body), ...rest});
 
@@ -1013,7 +1015,7 @@ language! {
         fold_proc(s.clone(), res) <--
             proc(s),
             if let Proc::CommWhere(ref pat, ref n, ref q, ref cond, ref body) = s,
-            let res = crate::for_clause::comm_pforwhere_subst(
+            let res = crate::rhocalc::receive::comm_pforwhere_subst(
                 pat.as_ref(),
                 n.as_ref(),
                 q.as_ref(),
@@ -1026,7 +1028,7 @@ language! {
         fold_proc(s.clone(), res) <--
             proc(s),
             if let Proc::PForUser(ref rows, ref body) = s,
-            let res = crate::for_clause::desugar_for_rows(rows.clone(), body.as_ref());
+            let res = crate::rhocalc::receive::desugar_for_rows(rows.clone(), body.as_ref());
 
         // many-step to a result
         relation path(Proc, Proc);
@@ -1089,294 +1091,4 @@ language! {
         //     proc(p),name(n),
         //     !(proc(k), trans(p,k,q), can_comm(q,n));
     },
-}
-
-fn infer_receive_pattern_names(pat: &Proc, out: &mut Vec<String>) {
-    match pat {
-        Proc::PVar(mettail_runtime::OrdVar(mettail_runtime::Var::Free(fv))) => {
-            if let Some(name) = &fv.pretty_name {
-                out.push(name.clone());
-            }
-        },
-        Proc::CastList(xs) => {
-            if let List::ListLit(items) = xs.as_ref() {
-                for item in items {
-                    infer_receive_pattern_names(item, out);
-                }
-            }
-        },
-        Proc::CastBag(xs) => {
-            if let Bag::BagLit(items) = xs.as_ref() {
-                for (item, count) in items.iter() {
-                    for _ in 0..count {
-                        infer_receive_pattern_names(item, out);
-                    }
-                }
-            }
-        },
-        Proc::CastMap(m) => {
-            if let Map::MapLit(items) = m.as_ref() {
-                for (_, value) in items.iter() {
-                    infer_receive_pattern_names(value, out);
-                }
-            }
-        },
-        _ => {},
-    }
-}
-
-fn name_uses_var(name: &Name, var_name: &str) -> bool {
-    match name {
-        Name::NVar(mettail_runtime::OrdVar(mettail_runtime::Var::Free(fv))) => {
-            fv.pretty_name.as_deref() == Some(var_name)
-        },
-        Name::NQuote(p) => proc_uses_name_var(p, var_name) || proc_uses_proc_var(p, var_name),
-        _ => false,
-    }
-}
-
-fn input_bind_uses_name_var(bind: &InputBind, var_name: &str) -> bool {
-    match bind {
-        InputBind::InputBind(pat, n) => {
-            proc_uses_name_var(pat, var_name) || name_uses_var(n, var_name)
-        },
-        _ => false,
-    }
-}
-
-fn input_bind_uses_proc_var(bind: &InputBind, var_name: &str) -> bool {
-    match bind {
-        InputBind::InputBind(pat, n) => {
-            proc_uses_proc_var(pat, var_name) || name_uses_var(n, var_name)
-        },
-        _ => false,
-    }
-}
-
-fn proc_uses_name_var(term: &Proc, var_name: &str) -> bool {
-    match term {
-        Proc::PPar(ps) => ps.iter().any(|(p, _)| proc_uses_name_var(p, var_name)),
-        Proc::POutput(n, q) => name_uses_var(n, var_name) || proc_uses_name_var(q, var_name),
-        Proc::PDrop(n) => name_uses_var(n, var_name),
-        Proc::PFor(_, n, body) => name_uses_var(n, var_name) || proc_uses_name_var(body, var_name),
-        Proc::PForWhere(_, n, cond, body) => {
-            name_uses_var(n, var_name)
-                || proc_uses_name_var(cond, var_name)
-                || proc_uses_name_var(body, var_name)
-        },
-        Proc::PForJoin(b, bs, cond, body) => {
-            input_bind_uses_name_var(b, var_name)
-                || bs.iter().any(|ib| input_bind_uses_name_var(ib, var_name))
-                || proc_uses_name_var(cond, var_name)
-                || proc_uses_name_var(body, var_name)
-        },
-        Proc::PForUser(rows, body) => {
-            let d = crate::for_clause::desugar_for_rows(rows.clone(), body);
-            proc_uses_name_var(&d, var_name)
-        },
-        Proc::GuardThen(cond, body) => {
-            proc_uses_name_var(cond, var_name) || proc_uses_name_var(body, var_name)
-        },
-        Proc::PNew(scope) => proc_uses_name_var(scope.unsafe_body(), var_name),
-        _ => false,
-    }
-}
-
-fn proc_uses_proc_var(term: &Proc, var_name: &str) -> bool {
-    match term {
-        Proc::PVar(mettail_runtime::OrdVar(mettail_runtime::Var::Free(fv))) => {
-            fv.pretty_name.as_deref() == Some(var_name)
-        },
-        Proc::PPar(ps) => ps.iter().any(|(p, _)| proc_uses_proc_var(p, var_name)),
-        Proc::POutput(n, q) => name_uses_var(n, var_name) || proc_uses_proc_var(q, var_name),
-        Proc::PDrop(n) => name_uses_var(n, var_name),
-        Proc::PFor(pat, n, body) => {
-            proc_uses_proc_var(pat, var_name)
-                || name_uses_var(n, var_name)
-                || proc_uses_proc_var(body, var_name)
-        },
-        Proc::PForWhere(pat, n, cond, body) => {
-            proc_uses_proc_var(pat, var_name)
-                || name_uses_var(n, var_name)
-                || proc_uses_proc_var(cond, var_name)
-                || proc_uses_proc_var(body, var_name)
-        },
-        Proc::PForJoin(b, bs, cond, body) => {
-            input_bind_uses_proc_var(b, var_name)
-                || bs.iter().any(|ib| input_bind_uses_proc_var(ib, var_name))
-                || proc_uses_proc_var(cond, var_name)
-                || proc_uses_proc_var(body, var_name)
-        },
-        Proc::PForUser(rows, body) => {
-            let d = crate::for_clause::desugar_for_rows(rows.clone(), body);
-            proc_uses_proc_var(&d, var_name)
-        },
-        Proc::GuardThen(cond, body) => {
-            proc_uses_proc_var(cond, var_name) || proc_uses_proc_var(body, var_name)
-        },
-        Proc::PNew(scope) => proc_uses_proc_var(scope.unsafe_body(), var_name),
-        _ => false,
-    }
-}
-
-fn infer_receive_var_type(body: &Proc, cond: Option<&Proc>, var_name: &str) -> TermType {
-    let uses_name =
-        proc_uses_name_var(body, var_name) || cond.is_some_and(|c| proc_uses_name_var(c, var_name));
-    let uses_proc =
-        proc_uses_proc_var(body, var_name) || cond.is_some_and(|c| proc_uses_proc_var(c, var_name));
-    if uses_name {
-        TermType::Base("Name".to_string())
-    } else if uses_proc {
-        TermType::Base("Proc".to_string())
-    } else {
-        TermType::Base("Name".to_string())
-    }
-}
-
-fn collect_rhocalc_var_types(
-    term: &Proc,
-    result: &mut Vec<VarTypeInfo>,
-    seen: &mut std::collections::HashSet<String>,
-) {
-    match term {
-        Proc::PForUser(rows, body) => {
-            let desugared = crate::for_clause::desugar_for_rows(rows.clone(), body);
-            collect_rhocalc_var_types(&desugared, result, seen);
-        },
-        Proc::PFor(pat, _n, body) => {
-            let mut names = Vec::new();
-            infer_receive_pattern_names(pat, &mut names);
-            for name in names {
-                if seen.insert(name.clone()) {
-                    result.push(VarTypeInfo {
-                        name: name.clone(),
-                        ty: infer_receive_var_type(body, None, &name),
-                    });
-                }
-            }
-            collect_rhocalc_var_types(body, result, seen);
-        },
-        Proc::PForWhere(pat, _n, cond, body) => {
-            let mut names = Vec::new();
-            infer_receive_pattern_names(pat, &mut names);
-            for name in names {
-                if seen.insert(name.clone()) {
-                    result.push(VarTypeInfo {
-                        name: name.clone(),
-                        ty: infer_receive_var_type(body, Some(cond), &name),
-                    });
-                }
-            }
-            collect_rhocalc_var_types(cond, result, seen);
-            collect_rhocalc_var_types(body, result, seen);
-        },
-        Proc::PForJoin(b, bs, cond, body) => {
-            let mut names = Vec::new();
-            if let InputBind::InputBind(pat, _) = b.as_ref() {
-                infer_receive_pattern_names(pat, &mut names)
-            }
-            for bind in bs {
-                if let InputBind::InputBind(pat, _) = bind {
-                    infer_receive_pattern_names(pat, &mut names)
-                }
-            }
-            for name in names {
-                if seen.insert(name.clone()) {
-                    result.push(VarTypeInfo {
-                        name: name.clone(),
-                        ty: infer_receive_var_type(body, Some(cond), &name),
-                    });
-                }
-            }
-            collect_rhocalc_var_types(cond, result, seen);
-            collect_rhocalc_var_types(body, result, seen);
-        },
-        Proc::PPar(ps) => {
-            for (p, _) in ps.iter() {
-                collect_rhocalc_var_types(p, result, seen);
-            }
-        },
-        Proc::GuardThen(cond, body) => {
-            collect_rhocalc_var_types(cond, result, seen);
-            collect_rhocalc_var_types(body, result, seen);
-        },
-        Proc::POutput(_, q) => collect_rhocalc_var_types(q, result, seen),
-        Proc::PNew(scope) => collect_rhocalc_var_types(scope.unsafe_body(), result, seen),
-        _ => {},
-    }
-}
-
-impl RhoCalcLanguage {
-    pub fn infer_var_types(&self, term: &dyn Term) -> Vec<VarTypeInfo> {
-        let Some(typed_term) = term.as_any().downcast_ref::<RhoCalcTerm>() else {
-            return <RhoCalcLanguage as Language>::infer_var_types(self, term);
-        };
-        match &typed_term.0 {
-            RhoCalcTermInner::Proc(p) => {
-                let mut result = Vec::new();
-                let mut seen = std::collections::HashSet::new();
-                collect_rhocalc_var_types(p, &mut result, &mut seen);
-                RhoCalcLanguage::collect_all_proc_vars(p, p, &mut result, &mut seen);
-                result
-            },
-            _ => <RhoCalcLanguage as Language>::infer_var_types(self, term),
-        }
-    }
-
-    pub fn infer_var_type(&self, term: &dyn Term, var_name: &str) -> Option<TermType> {
-        let Some(typed_term) = term.as_any().downcast_ref::<RhoCalcTerm>() else {
-            return <RhoCalcLanguage as Language>::infer_var_type(self, term, var_name);
-        };
-        if let RhoCalcTermInner::Proc(proc) = &typed_term.0 {
-            let desugared = match proc {
-                Proc::PForUser(rows, body) => {
-                    Some(crate::for_clause::desugar_for_rows(rows.clone(), body))
-                },
-                _ => None,
-            };
-            let proc = desugared.as_ref().unwrap_or(proc);
-            match proc {
-                Proc::PFor(pat, _n, body) => {
-                    let mut names = Vec::new();
-                    infer_receive_pattern_names(pat, &mut names);
-                    if names.iter().any(|n| n == var_name) {
-                        return Some(infer_receive_var_type(body, None, var_name));
-                    }
-                },
-                Proc::PForWhere(pat, _n, cond, body) => {
-                    let mut names = Vec::new();
-                    infer_receive_pattern_names(pat, &mut names);
-                    if names.iter().any(|n| n == var_name) {
-                        return Some(infer_receive_var_type(body, Some(cond), var_name));
-                    }
-                },
-                Proc::PForJoin(b, bs, cond, body) => {
-                    let mut names = Vec::new();
-                    if let InputBind::InputBind(pat, _) = b.as_ref() {
-                        infer_receive_pattern_names(pat, &mut names)
-                    }
-                    for bind in bs {
-                        if let InputBind::InputBind(pat, _) = bind {
-                            infer_receive_pattern_names(pat, &mut names)
-                        }
-                    }
-                    if names.iter().any(|n| n == var_name) {
-                        return Some(infer_receive_var_type(body, Some(cond), var_name));
-                    }
-                },
-                _ => {},
-            }
-            if let Some(t) = proc.infer_var_type(var_name) {
-                return Some(RhoCalcLanguage::inferred_to_term_type(&t));
-            }
-            let mut result = Vec::new();
-            let mut seen = std::collections::HashSet::new();
-            RhoCalcLanguage::collect_all_proc_vars(proc, proc, &mut result, &mut seen);
-            return result
-                .into_iter()
-                .find(|v| v.name == var_name)
-                .map(|v| v.ty);
-        }
-        <RhoCalcLanguage as Language>::infer_var_type(self, term, var_name)
-    }
 }
